requirements.txt
-i https://pypi.python.org/simple
gradio==3.*
iso639-lang
pycld2 # In future, maybe use lingua-language-detector
torch
transformers
sentencepiece
langdetect
sacremoses